# apply loop unroll trick