@gunnarmorling 2024-01-29
Nice!

```
Benchmark 1: timeout -v 300 ./calculate_average_abeobk.sh 2>&1
  Time (mean ± σ):      1.964 s ±  0.016 s    [User: 0.002 s, System: 0.003 s]
  Range (min … max):    1.935 s …  1.973 s    5 runs

Summary
  abeobk: trimmed mean 1.9700661416200003, raw times 1.93466153562,1.9687762126200001,1.9683303976200002,1.97309181462,1.9733764866200003

Leaderboard

| # | Result (m:s.ms) | Implementation     | JDK | Submitter     | Notes     |
|---|-----------------|--------------------|-----|---------------|-----------|
|   | 00:01.970 | [link](https://github.com/gunnarmorling/1brc/blob/main/src/main/java/dev/morling/onebrc/CalculateAverage_abeobk.java)| 21.0.2-graal | [Van Phu DO](https://github.com/abeobk) | GraalVM native binary, uses Unsafe |
```